requirements.txt
asqlite==2.0.0
discord.py==2.6.4
elevenlabs==2.26.0
fastapi==0.124.0
local-ffmpeg==0.1.4
gspread==6.2.1
keyboard==0.13.5
mutagen==1.47.0
numpy==2.3.5
oauth2client==4.1.3
obsws_python==1.8.0
openai==2.9.0
Pillow==12.0.0
pydub==0.25.1
pygame==2.6.1
python-dotenv==1.2.1
soundfile==0.13.1
tiktoken==0.11.0
twitchio==3.1.0
uvicorn==0.38.0
